#009 rendering setup (vulkan) (#6)
* Initial commit. Added Vulkan to project for Win32, pending additional import code for Mac and Linux. Fixed a minor typo in windowing code to ensure proper overlapped window status.

* Began work on Vulkan Renderer. Made restructuring modifications to Application. Finished Instance creation (without validation layers) in VulkanInit.

* All Vulkan Setup up to Swapchain complete.

* Added Assets folder for shaders. Will be replaced with compiled archive later. Added temporary shaders to facilitate testing. Completed graphics pipeline code.

* Renamed VkInit to VkCore. Finished initial version of draw code and achieved the first triangle. No errors reported, app opens for 5 seconds and closes as intended.

* Added Swapchain reset to render pipeline. Does not handle window resize or minimization, as the windows do not currently process those events. Will be added at a later date when more of Vulkan is complete.

* Swapchain recreation complete including window events for resizing and handling of minimization.

* Started work on Vertex Buffer rendering. Currently does not run, as there are no vertices in the vertex buffer.

* Rendering triangle with Vertex Buffer.

* Completed basic 3D rendering for setup code with a Uniform Buffer. Currently still hardcoded.

* Basic texture rendering working. Moving on to depth buffer.

* Depth buffer not working. Pushed code for Cameron.

* Fixed depth buffer rendering in the wrong order (accidental flipped z-axis on projection matrix)

* Slightly untangled buggy include behavior within renderer code.

* Continued work on separating Vulkan from the hard coded data. Began work on Render Contexts.

* Commit because I will have no internet access for a day and still want to have my work backed up --O

* Completed basic Uniform Buffer management, however, descriptors are currently not written.

* AAAlmost done, but a linker error is causing me pain. I'm going to take a break for now and push this.

* Busted code.

* Fixed linker error -- Thanks Robert!

* Verified working render contexts and cleaned up/refactored render functions. Render functions were split between BeginRenderPass, Draw, and Present. There are currently a few memory leaks within the Vulkan renderer that will be resolved when Uniform Buffer binding confirmed working.

* All rendering code confirmed working, images now properly render to the screen with pipelines now inside of modular data structures.

+++ Athena/src/Athena/Core/TestBlock.hpp
@@ -14,7 +14,7 @@ private:
 	size_t m_sFastestFuncNdx = 0;
 	size_t m_sSlowestFuncNdx = 0;
 
-	uint32_t m_u32SuccessfulTests = 0u;
+	uint64_t m_u32SuccessfulTests = 0u;
 public:
 	TestBlock() : m_strBlockName("DEFAULT"), m_vTests() {}
 	TestBlock(const std::string& _strName) : m_strBlockName(_strName), m_vTests() {}
@@ -35,6 +35,6 @@ public:
 	[[nodiscard]] HC_INLINE TestCase* GetFastestTestCase() { return m_vTests.size() > 0 ? &m_vTests[m_sFastestFuncNdx] : nullptr; }
 	[[nodiscard]] HC_INLINE TestCase* GetSlowestTestCase() { return m_vTests.size() > 0 ? &m_vTests[m_sSlowestFuncNdx] : nullptr; }
 
-	[[nodiscard]] HC_INLINE uint32_t GetSuccessfulTestCount() const { return m_u32SuccessfulTests; }
-	[[nodiscard]] HC_INLINE uint32_t GetTestCount() const { return m_vTests.size(); }
+	[[nodiscard]] HC_INLINE uint64_t GetSuccessfulTestCount() const { return m_u32SuccessfulTests; }
+	[[nodiscard]] HC_INLINE uint64_t GetTestCount() const { return m_vTests.size(); }
 };
